apps-rps/rps-game-agent/requirements.txt
flask==3.0.0
requests==2.31.0
azure-identity>=1.15.0
azure-ai-projects
openai>=1.0.0
python-dotenv>=1.0.0